rename gem and bump to version 0.9.0
