+++ lib/lightstep/global_tracer.rb
