+++ lib/lightstep/span.rb
