--- lib/lightstep/tracer/version.rb
@@ -1,5 +1,0 @@
-module LightStep
-  class Tracer
-    VERSION = '0.8.14'.freeze
-  end
-end
