+++ lib/lightstep/transport/base.rb
