+++ lib/lightstep/transport/nil.rb
@@ -1,4 +1,4 @@
-require 'lightstep/tracer/transport/base'
+require 'lightstep/transport/base'
 
 module LightStep
   module Transport
