+++ lib/lightstep/version.rb
@@ -1,0 +1,3 @@
+module LightStep
+  VERSION = '0.9.0'.freeze
+end
