+++ scripts/version.rb
@@ -1,5 +1,5 @@
 lib = File.expand_path('../../lib', __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'lightstep/tracer/version'
+require 'lightstep/version'
 
-print LightStep::Tracer::VERSION
+print LightStep::VERSION
